Modify get_random_state in n_step_rnn
